fix: break agent cycle, fix lint errors, update eslint config for renamed file
