+++ turbo/apps/platform/eslint.config.js
@@ -123,7 +123,7 @@ export default [
       "src/signals/zero-page/chat-draft.ts",
       "src/signals/__tests__/fetch.test.ts",
       "src/views/zero-page/components/org-manage/org-general-tab.tsx",
-      "src/views/zero-page/zero-jobs-page.tsx",
+      "src/views/agents-page/agents-page.tsx",
       "src/views/zero-page/zero-settings-tab.tsx",
     ],
     rules: {
